+++ src/models/index.ts
@@ -540,7 +540,6 @@ export * from "./restcollectorconf.js";
 export * from "./retryrulestype.js";
 export * from "./retryrulestype1.js";
 export * from "./retrysettingstype.js";
-export * from "./retrytypeoptionshealthcheckcollectorconfretryrules.js";
 export * from "./rollbacksettingsunion.js";
 export * from "./routeconf.js";
 export * from "./routes.js";